+++ NEWS.md
@@ -6,6 +6,7 @@
 * Python: added `arkimet.Summary.read_json` (#230)
 * Python: added `annotate: bool = False` to `arkimet.Summary.write` (#230)
 * Python: added `annotate: bool = False` to `arkimet.Summary.write_short` (#230)
+* Python: implemented `arkimet.Metadata.write` for `yaml` and `json` formats (#230)
 
 # New in version 1.27